requirements.txt
python-telegram-bot==21.11.1
sqlalchemy==2.0.40
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
sentry-sdk>=2.35.0
psycopg2-binary==2.9.9
pymysql==1.1.0
gunicorn==21.2.0

